+++ lisp/eab-ace.el
@@ -1,5 +1,4 @@
 (require 'ace-jump-mode)
-(require 'avy)
 ;; (require 'ace-window)
 
 (setq ace-jump-mode-gray-background nil) ;; it is faster