+++ lisp/eab-ui.el
@@ -2,6 +2,7 @@
 
 (eab/bind-path source-directory)
 (setq find-function-C-source-directory source-directory)
+(setq browse-url-browser-function (quote browse-url-firefox))
 
 
 ;; TODO depend on auto-complete, eab/trans-path
@@ -23,5 +24,28 @@
   (let ((str (filter-buffer-substring start end)))
     (kill-region start end)
     (insert (eab/latinize str))))
+
+;; DONE
+;; xset dpms force off/on
+;; gnome-screensaver-command -a
+
+(defun eab/screen-off ()
+  (interactive)
+  (sleep-for 0.2)
+  (eab/shell-command "xset dpms force off" nil 0)
+  (suspend-frame))
+
+(defun eab/gnome-terminal ()
+  (eab/shell-command "gnome-terminal" nil 0))
+
+(defun google ()
+  "Google the selected region if any, display a query prompt otherwise."
+  (interactive)
+  (browse-url
+   (concat
+    "http://www.google.com/search?ie=utf-8&oe=utf-8&q="
+    (url-hexify-string (if mark-active
+			   (buffer-substring (region-beginning) (region-end))
+			 (read-string "Google: "))))))
 
 (provide 'eab-ui)